# contracts/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Procura S2C API
  description: API de la plateforme Source-to-Contract Procura
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /auth/login:
    post:
      tags:
        - auth
      description: Authentifier l'utilisateur localement et retourner les tokens d'accès
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 6
      responses:
        "200":
          description: Authentification réussie
          content:
            application/json:
              schema:
                description: Authentification réussie
                type: object
                properties:
                  token:
                    type: string
                  refreshToken:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        format: uuid
                      displayName:
                        type: string
                      role:
                        type: string
                      email:
                        type: string
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string
  /auth/logout:
    post:
      tags:
        - auth
      description: Invalider la session active
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
...
servers:
  - url: http://0.0.0.0:8080
    description: Serveur de développement local